+++ magda-web-client/src/api-clients/OrgUnitApis.ts
@@ -27,3 +27,37 @@ export async function listOrgUnits({
         return await res.json();
     }
 }
+
+export interface OrgUnitWithRelationship extends OrgUnit {
+    relationship?: string;
+}
+
+/**
+ * List all OrgUnits at certain org tree level.
+ * Optionally provide a test Org Unit Id that will be used to
+ * test the relationship with each of returned orgUnit item.
+ * Possible Value: 'ancestor', 'descendant', 'equal', 'unrelated'.
+ *
+ * @export
+ * @param {number} orgLevel - The level number (starts from 1) where org Units of the tree are taken horizontally.
+ * @param {string} [relationshipOrgUnitId] - Optional; The org unit id that is used to test the relationship with each of returned orgUnit item.
+ * @returns {Promise<OrgUnitWithRelationship[]>}
+ */
+export async function listOrgUnitsAtLevel(
+    orgLevel: number,
+    relationshipOrgUnitId?: string
+): Promise<OrgUnitWithRelationship[]> {
+    const uri = `${config.authApiUrl}orgunits/bylevel/${orgLevel}${
+        relationshipOrgUnitId
+            ? `?relationshipOrgUnitId=${relationshipOrgUnitId}`
+            : ""
+    }`;
+
+    const res = await fetch(uri, config.fetchOptions);
+
+    if (!res.ok) {
+        throw new Error("Rejected with " + res.statusText);
+    } else {
+        return await res.json();
+    }
+}