+++ CHANGES.md
@@ -87,6 +87,7 @@ UI:
 -   Show an error message screen if the user is not allowed to access the add dataset page
 -   Updated security classification & sensitivity questions according to the new design
 -   Make sure all publish new dataset errors are captured and shown to users on add dataset page
+-   Added Custodian field to the 'People and Production' page
 -   Added the ability to add references to datasets that a new dataset was derived from.
 
 Gateway: